feat(safety): granular YAML control for sheets chart and named-ranges
Extract SheetsChartCmd and SheetsNamedRangesCmd into *_types.go so
the generator recognizes their subcommands as YAML keys. Profiles
can now permit read-only chart/named-ranges access without granting
the destructive subcommands.

readonly.yaml now exposes sheets.chart.list/get and
sheets.named-ranges.list/get (previously the entire subtree was
gated on a single boolean and excluded from readonly entirely).

agent-safe.yaml enables all chart subcommands (recoverable via
revision history) and keeps named-ranges writes off (still
conservative about formula reference changes).

Verified by smoke testing each profile:
- readonly: gog sheets chart shows only list/get
- agent-safe: gog sheets chart shows list/get/create/update/delete
- agent-safe and readonly: gog gmail forward absent (unchanged)
